backend/requirements.txt
python-dotenv
supabase
flask
flask-cors
openai>=1.0.0